API/requirements.txt
fastapi
uvicorn[standard]
pydantic
SQLAlchemy
python-dotenv
